+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0
+
+- [Export resources created by this module](https://github.com/babbel/terraform-aws-athena/pull/2)
+- [Remove policy document `iam-policy-document-athena-workspace-use` output](https://github.com/babbel/terraform-aws-athena/pull/2)
+
 ## v1.0.0
 
 - [Initial version](https://github.com/babbel/terraform-aws-athena/pull/1)